refactor: centralize validated contract descriptors
