+++ core/foundation/c2-contract/src/release.rs
@@ -1,0 +1,32 @@
+use crate::{ContractError, validate_contract_hash};
+use std::fmt;
+use std::str::FromStr;
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct ContractDescriptorDigest(String);
+
+impl ContractDescriptorDigest {
+    pub fn parse(value: impl Into<String>) -> Result<Self, ContractError> {
+        let value = value.into();
+        validate_contract_hash("descriptor_sha256", &value)?;
+        Ok(Self(value))
+    }
+
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+}
+
+impl fmt::Display for ContractDescriptorDigest {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        formatter.write_str(self.as_str())
+    }
+}
+
+impl FromStr for ContractDescriptorDigest {
+    type Err = ContractError;
+
+    fn from_str(value: &str) -> Result<Self, Self::Err> {
+        Self::parse(value)
+    }
+}
