+++ tests/fixtures/contracts/portable-release.canonical.json
@@ -1,0 +1,1 @@
+{"crm":{"name":"Portable","namespace":"test.contract-release","version":"0.1.0"},"fingerprints":{"abi_hash":"0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef","signature_hash":"abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789"},"methods":[{"access":"read","buffer":"view","name":"ping","parameters":[],"return":{"kind":"none"},"wire":{"input":null,"output":null}},{"access":"write","buffer":"view","name":"echo","parameters":[{"default":{"kind":"missing"},"kind":"POSITIONAL_OR_KEYWORD","name":"value","type":{"codec":{"capabilities":["bytes","buffer-view"],"id":"org.fastdb.call-db","kind":"codec_ref","portable":true,"schema":"fastdb.call-db.schema.v1","schema_sha256":"aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa","version":"1"},"kind":"codec"}}],"return":{"codec":{"capabilities":["bytes","buffer-view"],"id":"org.fastdb.call-db","kind":"codec_ref","portable":true,"schema":"fastdb.call-db.schema.v1","schema_sha256":"aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa","version":"1"},"kind":"codec"},"wire":{"input":{"capabilities":["bytes","buffer-view"],"id":"org.fastdb.call-db","kind":"codec_ref","portable":true,"schema":"fastdb.call-db.schema.v1","schema_sha256":"aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa","version":"1"},"output":{"capabilities":["bytes","buffer-view"],"id":"org.fastdb.call-db","kind":"codec_ref","portable":true,"schema":"fastdb.call-db.schema.v1","schema_sha256":"aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa","version":"1"}}}],"schema":"c-two.contract.v1"}
